requirements.txt
flask
firebase-admin
flask-cors
python-dotenv
Flask-Mail
gunicorn
